refactor(proxy): 抽离代理请求生命周期 module
